Fix chemical, cleanup
