+++ package.json
@@ -1,26 +1,16 @@
 {
-  "name": "proxython2024",
+  "name": "fluid",
   "version": "1.0.0",
   "description": "",
-  "main": "server.js",
+  "main": "index.js",
   "scripts": {
-    "start": "node fix.js"
+    "start": "node index.js"
   },
   "author": "NC & Peak",
   "type": "module",
   "dependencies": {
-    "@mercuryworkshop/bare-as-module3": "^2.2.2",
-    "@mercuryworkshop/bare-mux": "^2.0.4",
-    "@mercuryworkshop/epoxy-transport": "^2.1.3",
-    "@mercuryworkshop/libcurl-transport": "^1.3.7",
-    "@mercuryworkshop/scramjet": "^1.0.2",
-    "@titaniumnetwork-dev/ultraviolet": "^3.2.6",
-    "@tomphttp/bare-server-node": "^2.0.3",
     "chemicaljs": "^1.6.2",
-    "compress": "^0.99.0",
     "compression": "^1.7.4",
-    "express": "^4.19.2",
-    "wisp": "^0.10.0",
-    "wisp-server-node": "^1.1.0"
+    "express": "^4.19.2"
   }
 }
