--- public/assets/js/scram/scramjet.config.js
@@ -1,9 +1,0 @@
-self.__scramjet$config = {
-  prefix: "/~/scram/",
-  codec: self.__scramjet$codecs.xor,
-  config: "/assets/js/scram/scramjet.config.js",
-  bundle: "/assets/js/scram/scramjet.bundle.js",
-  worker: "/assets/js/scram/scramjet.worker.js",
-  client: "/assets/js/scram/scramjet.client.js",
-  codecs: "/assets/js/scram/scramjet.codecs.js",
-};
