+++ public/index.html
@@ -26,11 +26,6 @@
     <link href="/assets/css/global.css" rel="stylesheet" type="text/css" />
     <link href="assets/css/themes.css" rel="stylesheet" type="text/css" />
     <link href="/favicon.ico" rel="icon" id="icon" />
-    <!-- <script src="/assets/js/uv/uv.bundle.js" defer></script>
-    <script src="/assets/js/uv/uv.config.js" defer></script>
-    <script src="/assets/js/scram/scramjet.codecs.js"></script>
-    <script src="/assets/js/scram/scramjet.bundle.js" defer></script>
-    <script src="/assets/js/scram/scramjet.config.js" defer></script> -->
 
     <script src="/chemical.js"></script>
     <script type="module" src="/assets/js/tabs_manager.js" defer></script>
